requirements.txt
# All your requirements go here. Please adapt as needed:
typer
pydantic
PyYAML
pytest
pytest-cov
pytest-asyncio
mypy
bandit
pylint
black
flake8
isort
pre-commit
testcontainers
kafka-python
